requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.0.3
pydantic-settings==2.0.3

langchain==0.0.267
langchain-community==0.0.10
langchain-text-splitters==0.0.1
langchain-openai==0.0.2
openai==1.3.0
sentence-transformers==2.2.2

pinecone-client==2.2.2
faiss-cpu==1.7.4

loguru==0.7.0
tqdm==4.66.1

python-multipart==0.0.9
jinja2==3.1.2

pytest==7.4.0
httpx==0.24.1

numpy==1.24.3
PyYAML==6.0

# 流式響應和異步支持
aiohttp==3.8.5
sse-starlette==1.6.5

# 雲部署支持
vercel==0.2.1